Создал permission, проверящее, что польльзователь является владельцем аккаунта или является суперпользователем

+++ accounts/views.py
@@ -5,6 +5,8 @@ from .models import CustomUser
 from rest_framework.response import Response
 from rest_framework import status
 
+from .permissions import IsUserAccountOwnerOrAdmin
+
 class UserCreate(APIView):
     serializer_class = CustomUserSerializer
 
@@ -19,13 +21,15 @@ class UserCreate(APIView):
 
 class UserDelete(APIView):
     serializer_class = CustomUserSerializer
+    permission_classes = (IsUserAccountOwnerOrAdmin,)
 
     def delete(self, request, pk):
         try:
             user_to_delete = CustomUser.objects.get(pk=pk)
         except CustomUser.DoesNotExist:
             return Response(status=status.HTTP_400_BAD_REQUEST, data="User not found!")
-
+        
+        self.check_object_permissions(self.request, user_to_delete)
         user_to_delete.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT, data="User deleted!")
